+++ app/views/two_factor_authentication/shared/max_login_attempts_reached.html.erb
@@ -5,7 +5,7 @@
   <%= t('titles.account_locked') %>
 </h1>
 <p>
-  <%= t('devise.two_factor_authentication.max_login_attempts_reached') %>
+<%= t("devise.two_factor_authentication.max_#{type}_login_attempts_reached") %>
 </p>
 <p>
   <%= t('devise.two_factor_authentication.please_try_again_html',
