+++ spec/features/visitors/password_recovery_spec.rb
@@ -1,43 +1,7 @@
 require 'rails_helper'
 
 feature 'Password Recovery' do
-  def reset_password_and_sign_back_in(user, password = 'a really long password')
-    fill_in t('forms.passwords.edit.labels.password'), with: password
-    click_button t('forms.passwords.edit.buttons.submit')
-    fill_in_credentials_and_submit(user.email, password)
-  end
-
-  def recovery_code_from_pii(user, pii)
-    profile = create(:profile, :active, :verified, user: user)
-    pii_attrs = Pii::Attributes.new_from_hash(pii)
-    user_access_key = user.unlock_user_access_key(user.password)
-    recovery_code = profile.encrypt_pii(user_access_key, pii_attrs)
-    profile.save!
-
-    recovery_code
-  end
-
-  def trigger_reset_password_and_click_email_link(email)
-    visit new_user_password_path
-    fill_in 'Email', with: email
-    click_button t('forms.buttons.continue')
-    open_last_email
-    click_email_link_matching(/reset_password_token/)
-  end
-
-  def scrape_recovery_code
-    new_recovery_code_words = []
-    page.all(:css, '[data-recovery]').each do |node|
-      new_recovery_code_words << node.text
-    end
-    new_recovery_code_words.join(' ')
-  end
-
-  def reactivate_profile(password, recovery_code)
-    fill_in 'Password', with: password
-    enter_recovery_code(code: recovery_code)
-    click_button t('forms.reactivate_profile.submit')
-  end
+  include RecoveryCodeHelper
 
   context 'user enters valid email in forgot password form', email: true do
     it 'redirects to forgot_password path and sends an email to the user' do
@@ -245,78 +209,6 @@ feature 'Password Recovery' do
 
         expect(page).to have_content 'is too short'
       end
-    end
-  end
-
-  context 'LOA3 user' do
-    let(:user) { create(:user, :signed_up) }
-    let(:new_password) { 'some really awesome new password' }
-    let(:pii) { { ssn: '666-66-1234', dob: '1920-01-01' } }
-
-    scenario 'resets password and reactivates profile with recovery code', email: true do
-      recovery_code = recovery_code_from_pii(user, pii)
-
-      trigger_reset_password_and_click_email_link(user.email)
-
-      reset_password_and_sign_back_in(user, new_password)
-      click_submit_default
-      enter_correct_otp_code_for_user(user)
-
-      expect(current_path).to eq reactivate_profile_path
-
-      reactivate_profile(new_password, recovery_code)
-
-      expect(page).to have_content t('idv.messages.recovery_code')
-    end
-
-    scenario 'resets password, makes recovery code, attempts reactivate profile', email: true do
-      _recovery_code = recovery_code_from_pii(user, pii)
-
-      trigger_reset_password_and_click_email_link(user.email)
-
-      reset_password_and_sign_back_in(user, new_password)
-      click_submit_default
-      enter_correct_otp_code_for_user(user)
-
-      expect(current_path).to eq reactivate_profile_path
-
-      visit manage_recovery_code_path
-
-      new_recovery_code = scrape_recovery_code
-      click_acknowledge_recovery_code
-
-      expect(current_path).to eq reactivate_profile_path
-
-      reactivate_profile(new_password, new_recovery_code)
-
-      expect(page).to have_content t('errors.messages.recovery_code_incorrect')
-    end
-
-    scenario 'resets password, uses recovery code as 2fa', email: true do
-      recovery_code = recovery_code_from_pii(user, pii)
-
-      trigger_reset_password_and_click_email_link(user.email)
-
-      reset_password_and_sign_back_in(user, new_password)
-      click_submit_default
-
-      click_link t('devise.two_factor_authentication.recovery_code_fallback.link')
-
-      enter_recovery_code(code: recovery_code)
-
-      click_submit_default
-
-      expect(current_path).to eq sign_up_recovery_code_path
-
-      new_recovery_code = scrape_recovery_code
-      click_acknowledge_recovery_code
-
-      expect(current_path).to eq reactivate_profile_path
-
-      reactivate_profile(new_password, new_recovery_code)
-
-      expect(page).to_not have_content t('errors.messages.recovery_code_incorrect')
-      expect(page).to have_content t('idv.messages.recovery_code')
     end
   end
